+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Improve displaying params of array type in sort box (roubkar)
+
 ## 1.3.2 Jan 22 2021
 - Resolve search issue of sort box (gorarakelyan)